Split all of the helper classes in closure glyph segmenter out.
Move them into their own files.
